moved reparent operations to vtorc
